requirements.txt
appdirs==1.4.3
bottle==0.12.18
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cheroot==8.2.1
cryptography==2.9
decorator==4.4.2
distlib==0.3.0
dnspython==1.16.0
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
future==0.18.2
idna==2.9
jaraco.functools==3.0.0
Markdown==3.1.1
mccabe==0.6.1
more-itertools==8.2.0
mysql-connector-python==8.0.21
protobuf==3.12.2
public==2019.4.13
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
pymongo==3.10.1
python-telegram-bot==12.6.1
query-string==2019.4.13
requests==2.23.0
six==1.14.0
tornado==6.0.4
urllib3==1.25.8
virtualenv==20.0.25
web.py==0.40
